+++ examples/statement-playground/index.ts
@@ -5,13 +5,13 @@ import {
   stringToTopic,
   type Statement,
 } from "@polkadot-api/sdk-statement"
-import { getWsProvider } from "@polkadot-api/ws-provider"
+import { getWsProvider } from "polkadot-api/ws"
 import { sign, getPublicKey } from "@scure/sr25519"
-import { fromHex, jsonPrint, mergeUint8 } from "@polkadot-api/utils"
 import { Binary, Blake2256 } from "@polkadot-api/substrate-bindings"
+import { jsonSerialize, mergeUint8 } from "polkadot-api/utils"
 
 // use any key
-const ALICE_SK = fromHex(process.env.ALICE_SK!)
+const ALICE_SK = Binary.fromHex(process.env.ALICE_SK!)
 const alice = getStatementSigner(getPublicKey(ALICE_SK), "sr25519", (p) =>
   sign(ALICE_SK, p, Blake2256(mergeUint8([ALICE_SK, p]))),
 )
@@ -33,15 +33,16 @@ const signed1 = await alice.sign(stmt1)
 console.log(await sdk.submit(signed1))
 
 console.log(
-  jsonPrint(
+  JSON.stringify(
     await sdk.getStatements({
       dest: stringToTopic("key"),
       topics: [stringToTopic("1"), stringToTopic("2")],
     }),
+    jsonSerialize,
   ),
 )
 
 // GET ALL STATEMENTS (i.e. `dump`)
-console.log(jsonPrint(await sdk.getStatements()))
+console.log(JSON.stringify(await sdk.getStatements(), jsonSerialize))
 
 client.destroy()