+++ packages/sdk-ink/src/encoding-provider.ts
@@ -4,6 +4,11 @@ import {
   getInkLookup,
 } from "@polkadot-api/ink-contracts"
 import {
+  compactNumber,
+  HexString,
+  SizedHex,
+} from "@polkadot-api/substrate-bindings"
+import {
   AbiConstructor,
   AbiError,
   AbiFallback,
@@ -11,7 +16,7 @@ import {
   AbiParameter,
   AbiReceive,
 } from "abitype"
-import { Binary, FixedSizeBinary } from "polkadot-api"
+import { Binary } from "polkadot-api"
 import {
   Abi,
   decodeErrorResult,
@@ -21,36 +26,30 @@ import {
   encodeFunctionData,
 } from "viem"
 import { GenericInkDescriptors } from "./descriptor-types"
-import { compactNumber } from "@polkadot-api/substrate-bindings"
 
 export interface EncodingProvider {
-  isCompatible(codeHash: FixedSizeBinary<32>): boolean
+  isCompatible(codeHash: SizedHex<32>): boolean
   message(message: string): {
-    encode: (value: unknown) => Binary
-    decode: (value: Binary) => unknown
+    encode: (value: unknown) => Uint8Array
+    decode: (value: Uint8Array) => unknown
   }
   constructor(constructor: string): {
-    encode: (value: unknown) => Binary
-    decode: (value: Binary) => unknown
+    encode: (value: unknown) => Uint8Array
+    decode: (value: Uint8Array) => unknown
   }
   filterEvents(
     address: string,
-    events?: Array<
-      | {
-          event: GenericEvent
-          topics: Binary[]
-        }
-      | (GenericEvent & {
-          topics: Binary[]
-        })
-    >,
+    events?: Array<{
+      event: GenericEvent
+      topics: HexString[]
+    }>,
   ): GenericEvent[]
   storage(key: string): {
-    encode: (value: unknown) => Binary
-    decode: (value: Binary) => unknown
+    encode: (value: unknown) => Uint8Array
+    decode: (value: Uint8Array) => unknown
   }
   storagePaths(): string[]
-  decodeError(data: Binary): unknown
+  decodeError(data: Uint8Array): unknown
 }
 
 export const inkEncoding = (
@@ -64,13 +63,13 @@ export const inkEncoding = (
 
   return {
     isCompatible(codeHash) {
-      return codeHash ? codeHash.asHex() === lookup.metadata.source.hash : false
+      return codeHash ? codeHash === lookup.metadata.source.hash : false
     },
     message(message) {
       const msg = inkClient.message(message)
 
       return {
-        encode: (value) => msg.encode(value as any),
+        encode: (value) => msg.encode(value),
         decode: (data) => msg.decode({ data }),
       }
     },
@@ -78,7 +77,7 @@ export const inkEncoding = (
       const msg = inkClient.constructor(constructor)
 
       return {
-        encode: (value) => msg.encode(value as any),
+        encode: (value) => msg.encode(value),
         decode: (data) => msg.decode({ data }),
       }
     },
@@ -89,7 +88,7 @@ export const inkEncoding = (
       const storage = inkClient.storage(key)
 
       return {
-        encode: (value) => storage.encode(value as any),
+        encode: (value) => storage.encode(value),
         decode: (data) => storage.decode(data),
       }
     },
@@ -101,14 +100,14 @@ export const inkEncoding = (
       // Meaning what we get over the wire is [payload_len,data]
       // And `data` is a Vec<u8>, hence [msg_len,...chars]. In this case, the msg_len is redundant.
       try {
-        const bytes = data.asBytes()
+        const bytes = data
         const length = compactNumber.dec(bytes)
         const compactLength = compactNumber.enc(length).length
         if (compactLength + length === bytes.length) {
-          return Binary.fromBytes(bytes.slice(compactLength)).asText()
+          return Binary.toText(bytes.slice(compactLength))
         }
       } catch {}
-      return data.asHex()
+      return data
     },
   }
 }
@@ -124,7 +123,7 @@ export const solEncoding = (
 
   // There are two objects that are different between papi and viem:
   // - function: { address: HexString, selector: HexString } <=> HexString
-  // - bytes: Binary <=> HexString
+  // - bytes: Uint8Array <=> HexString
   const valueToViem = (value: unknown, param: AbiParameter): unknown => {
     if (param.type === "function") {
       if (typeof value === "string" && value.startsWith("0x")) return value
@@ -166,7 +165,7 @@ export const solEncoding = (
           "Expected Binary in parameter " + (param.name ?? "(no name)"),
         )
       }
-      return value.asHex()
+      return value
     }
 
     return value
@@ -315,34 +314,32 @@ export const solEncoding = (
           const result = decodeFunctionResult({
             abi,
             functionName: message === "fallback" ? undefined : message,
-            data: value.asHex() as `0x${string}`,
+            data: Binary.toHex(value) as `0x${string}`,
           })
           return viemToOutputs(result, "outputs" in msg ? msg.outputs : [])
         },
       }
     },
     filterEvents(address, events = []) {
-      const addrEq = (a: string | Binary) =>
-        (a instanceof Binary ? a.asHex() : a) === address
+      const addrEq = (a: string | Uint8Array) =>
+        (a instanceof Uint8Array ? Binary.toHex(a) : a) === address
 
-      const contractEvents = events
-        .map((v) => ("event" in v ? v : { event: v, topics: v.topics }))
-        .filter(
-          (v) =>
-            v.event.type === "Revive" &&
-            (v.event.value as any).type === "ContractEmitted" &&
-            addrEq((v.event.value as any).value.contract),
-        )
+      const contractEvents = events.filter(
+        (v) =>
+          v.event.type === "Revive" &&
+          (v.event.value as any).type === "ContractEmitted" &&
+          addrEq((v.event.value as any).value.contract),
+      )
 
       return contractEvents
         .map((v): GenericEvent | null => {
-          const topics = v.topics.map((v) => v.asHex())
+          const topics = v.topics.map((v) => v)
 
           try {
             const evtLog = decodeEventLog({
               abi,
               topics: topics as any,
-              data: (v.event.value as any).value.data.asHex(),
+              data: (v.event.value as any).value.data,
             })
             return {
               type: evtLog.eventName!,
@@ -364,7 +361,7 @@ export const solEncoding = (
       try {
         const error = decodeErrorResult({
           abi,
-          data: data.asHex() as `0x${string}`,
+          data: Binary.toHex(data) as `0x${string}`,
         })
         const abiItem = error.abiItem as AbiError
 
@@ -374,7 +371,7 @@ export const solEncoding = (
         }
       } catch (ex) {
         console.error(ex)
-        return data.asHex()
+        return data
       }
     },
   }