+++ packages/sdk-ink/src/util.ts
@@ -1,15 +1,18 @@
 import { RLP } from "@ethereumjs/rlp"
-import { Keccak256 } from "@polkadot-api/substrate-bindings"
+import {
+  ethAccount,
+  Keccak256,
+  SizedHex,
+} from "@polkadot-api/substrate-bindings"
 import {
   AccountId,
   Binary,
   Enum,
-  FixedSizeBinary,
+  FixedSizeArray,
   HexString,
   SS58String,
 } from "polkadot-api"
 import { mergeUint8 } from "polkadot-api/utils"
-import { ReviveAddress, ReviveSdkTypedApi, U256 } from "./descriptor-types"
 import { CommonTypedApi } from "./sdk-types"
 
 export const getSignedStorage = (
@@ -29,20 +32,19 @@ export const getStorageLimit = (
   }>,
 ) => (depositResponse.type === "Charge" ? depositResponse.value : 0n)
 
-export const ss58ToEthereum = (address: SS58String): Binary =>
-  Binary.fromBytes(Keccak256(AccountId().enc(address)).slice(12))
+export const ss58ToEthereum = (address: SS58String): SizedHex<20> =>
+  ethAccount.dec(Keccak256(AccountId().enc(address)).slice(12))
 
-/**
- * @deprecated Use `createInkSdk(client).addressIsMapped(address)` instead.
- */
 export const reviveAddressIsMapped = (
-  typedApi: ReviveSdkTypedApi | CommonTypedApi,
+  typedApi: CommonTypedApi,
   address: SS58String,
 ) =>
   typedApi.query.Revive.OriginalAccount.getValue(ss58ToEthereum(address)).then(
     (r) => r != null,
   )
 
+// TODO update to new format
+export type U256 = FixedSizeArray<4, bigint>
 const u64Range = 2n ** 64n
 export const valueToU256 = (value: bigint, nativeToEth: number): U256 => {
   const scaled = value * BigInt(nativeToEth)
@@ -61,45 +63,41 @@ export const u256ToValue = (u256: U256, nativeToEth: number): bigint => {
   return scaled / BigInt(nativeToEth)
 }
 
-const parseReviveAddress = (address: SS58String | ReviveAddress | HexString) =>
-  typeof address === "string"
-    ? address.startsWith("0x")
-      ? Binary.fromHex(address)
-      : ss58ToEthereum(address)
-    : address
+const parseReviveAddress = (address: SS58String | SizedHex<20>) =>
+  Binary.fromHex(address.startsWith("0x") ? address : ss58ToEthereum(address))
 
 // Ported from https://github.com/paritytech/polkadot-sdk/blob/c5ae50f86b8b727428eb86d9b1027a8f56fee19d/substrate/frame/revive/src/address.rs#L233
 export const getDeploymentAddressWithNonce = (
-  deployer: SS58String | ReviveAddress | HexString,
+  deployer: SS58String | SizedHex<20>,
   nonce: number,
 ) => {
   const addr = parseReviveAddress(deployer)
-  const data = RLP.encode([addr.asBytes(), nonce])
+  const data = RLP.encode([addr, nonce])
   const bytes = Keccak256(data).slice(12)
-  return Binary.fromBytes(bytes)
+  return Binary.toHex(bytes)
 }
 
 // Ported from https://github.com/paritytech/polkadot-sdk/blob/c5ae50f86b8b727428eb86d9b1027a8f56fee19d/substrate/frame/revive/src/address.rs#L242
 export const getDeploymentAddressWithSalt = (
-  deployer: SS58String | ReviveAddress | HexString,
-  deploymentHash: HexString,
-  salt: HexString | FixedSizeBinary<32>,
+  deployer: SS58String | SizedHex<20>,
+  deploymentHash: SizedHex<32>,
+  salt: SizedHex<32>,
 ) => {
   const addr = parseReviveAddress(deployer)
   const saltBin = typeof salt === "string" ? Binary.fromHex(salt) : salt
   const bytes = Keccak256(
     mergeUint8([
       new Uint8Array([0xff]),
-      addr.asBytes(),
-      saltBin.asBytes(),
-      Binary.fromHex(deploymentHash).asBytes(),
+      addr,
+      saltBin,
+      Binary.fromHex(deploymentHash),
     ]),
   ).slice(12)
 
-  return Binary.fromBytes(bytes)
+  return Binary.toHex(bytes)
 }
 
-export const getDeploymentHash = (code: Binary, inputData: Binary): HexString =>
-  Binary.fromBytes(
-    Keccak256(mergeUint8(code.asBytes(), inputData.asBytes())),
-  ).asHex()
+export const getDeploymentHash = (
+  code: Uint8Array,
+  inputData: Uint8Array,
+): HexString => Binary.toHex(Keccak256(mergeUint8([code, inputData])))