+++ packages/sdk-statement/src/codec/codec.ts
@@ -1,10 +1,11 @@
 import {
-  Bin,
-  Binary,
+  Bytes,
   CodecType,
   enhanceCodec,
   Enum,
-  FixedSizeBinary,
+  HexString,
+  SizedBytes,
+  SizedHex,
   Struct,
   u32,
   u64,
@@ -13,23 +14,23 @@ import {
 } from "@polkadot-api/substrate-bindings"
 
 export type Proof = Enum<{
-  sr25519: { signature: FixedSizeBinary<64>; signer: FixedSizeBinary<32> }
-  ed25519: { signature: FixedSizeBinary<64>; signer: FixedSizeBinary<32> }
-  ecdsa: { signature: FixedSizeBinary<65>; signer: FixedSizeBinary<33> }
+  sr25519: { signature: SizedHex<64>; signer: SizedHex<32> }
+  ed25519: { signature: SizedHex<64>; signer: SizedHex<32> }
+  ecdsa: { signature: SizedHex<65>; signer: SizedHex<33> }
   onChain: {
-    who: FixedSizeBinary<32>
-    blockHash: FixedSizeBinary<32>
+    who: SizedHex<32>
+    blockHash: SizedHex<32>
     event: bigint
   }
 }>
 
 export type Statement = Partial<{
   proof: Proof
-  decryptionKey: FixedSizeBinary<32>
+  decryptionKey: SizedHex<32>
   priority: number
-  channel: FixedSizeBinary<32>
-  topics: Array<FixedSizeBinary<32>>
-  data: Binary
+  channel: SizedHex<32>
+  topics: Array<SizedHex<32>>
+  data: Uint8Array
 }>
 
 const sortIdxs = {
@@ -48,14 +49,14 @@ const sortIdxs = {
 export type UnsignedStatement = Omit<Statement, "proof">
 export type SignedStatement = UnsignedStatement & { proof: Proof }
 
-const bin32 = Bin(32)
-const bin64 = Bin(64)
+const bin32 = SizedBytes(32)
+const bin64 = SizedBytes(64)
 
 const field = Variant({
   proof: Variant({
     sr25519: Struct({ signature: bin64, signer: bin32 }),
     ed25519: Struct({ signature: bin64, signer: bin32 }),
-    ecdsa: Struct({ signature: Bin(65), signer: Bin(33) }),
+    ecdsa: Struct({ signature: SizedBytes(65), signer: SizedBytes(33) }),
     onChain: Struct({ who: bin32, blockHash: bin32, event: u64 }),
   }),
   decryptionKey: bin32,
@@ -65,7 +66,7 @@ const field = Variant({
   topic2: bin32,
   topic3: bin32,
   topic4: bin32,
-  data: Bin(),
+  data: Bytes(),
 })
 const innerStatement = Vector(field)
 
@@ -87,7 +88,9 @@ export const statementCodec = enhanceCodec<
           stmt[k]!.forEach((v, i) => {
             statement.push(Enum(`topic${i + 1}` as `topic${1 | 2 | 3 | 4}`, v))
           })
-        } else statement.push(Enum(k, stmt[k]!))
+        } else {
+          statement.push(Enum(k, stmt[k]!))
+        }
       })
     return statement
   },
@@ -107,7 +110,7 @@ export const statementCodec = enhanceCodec<
         throw new Error(`Unexpected ${v.type}`)
       } else {
         statement.topics ??= []
-        statement.topics?.push(v.value as Binary)
+        statement.topics?.push(v.value as HexString)
       }
     })
     return statement