+++ packages/common-utils/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- BREAKING!: Update to PAPI v2
+- AsyncTransaction `getEncodedData` and `getBareTx` return a `Uint8Array`
+
 ## 0.2.0 2025-11-28
 
 ### Fixed